Backend/requirements.txt
psycopg2-binary>=2.9.9
# Bumped to satisfy crewai (requires >=1.1.1)
python-dotenv>=1.1.1
flask==2.3.3
flask-cors==4.0.0
flask-session>=0.5.0
werkzeug>=2.3.0
openai-whisper>=20250625
pyttsx3==2.90
pydub==0.25.1
sentence-transformers>=5.1.1
huggingface_hub>=0.35.1
scikit-learn>=1.4.0
openai>=1.0.0
numpy>=1.26.0
PyPDF2>=3.0.0
python-docx>=1.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
crewai>=0.203.0
langchain>=0.3.0
langchain-openai>=0.2.0
requests>=2.31.0